Renames lib/mongodb-fs to lib/server.
